Export fix for large datasers

+++ grails-app/domain/org/transmartproject/db/i2b2data/ObservationFact.groovy
@@ -33,7 +33,7 @@ class ObservationFact implements Serializable {
 
     String     valueType
     String     textValue
-    BigDecimal numberValue
+    Double numberValue
     String     valueFlag
     String     sourcesystemCd
